Add QQuickStyle::settings()
Loads ":/qtlabscontrols.conf" from resources. The next step is to make it
possible to specify the preferred theme and accent in qtlabscontrols.conf.
Different values can be specified for each style, which will take effect
when running the app using the respective style. These are the preferred
or default values, which can be overridden in code.

    [Controls]
    Style=material

    [Universal]
    Theme=Dark
    Accent=Red

    [Material]
    Theme=Light
    Accent=Brown

The reason for using our own qtlabscontrols.conf file instead of re-using
qt.conf is that qt.conf is strictly purposed for specifying the standard
Qt paths for deployment setups. If qt.conf missing paths for QLibraryInfo,
Qt fails to load the platform plugin:

    This application failed to start because it could not find
    or load the Qt platform plugin "windows|cocoa|xcb".

Change-Id: I8da7f0859f004db8adf585b830bce4aa8e7713a9
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>