V4 Debugger: use engineAdded() instead of engineAboutToBeAdded()
It does not make a difference in functionality, but after engineAdded()
the server won't wait on a mutex anymore. Before this change, if you
managed to send a message to the V4 debugger after the server had
called aboutToBeAdded(), but before it had stopped waiting, you could
produce a deadlock by scheduling an event for the GUI thread that was
never delivered.

Change-Id: I3a7fb2e46d67b7cc03733154ea61f2bf35f8ea24
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>