V4 Debugger: Clean up tests
Now that we can evaluate expressions without pausing, we might as well
test for that. The race conditions should also be resolved by now.

Clarify the reason why the break-on-attach test cannot work (so that
whoever tries to fix it the next time doesn't have to figure it out
for the (n+1)th time) and remove some dead code.

Change-Id: Ib1aab03d40779c6876db914b852f05061a2be046
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>