Fix ScrollBar & ScrollIndicator for the new Gallery
The Gallery example exposed some new bugs. We must let the scroll
bar/indicator initialize its size before layouting it, or else it
will be posioned on the edge and eventually grows outside of the
flickable it is attached to. Therefore, we must monitor the size
changes of the scroll bar/indicator the same way we're monitoring
the size changes of the attached flickable. Furthermore, while
debugging the issue, I noticed that QQuickScrollBar/Indicator were
unnecessarily monitoring all geometry changes. Monitoring only size
changes is enough.

Change-Id: I2581dba29bb4606642ba470dce85534632d7752e
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>