Drawer: tweak thresholds & offsets
This makes the drawer behave much better. Use increased threshold
only for dragging, not for presses (steals too large area from the
side). When closing by dragging from the outside of the drawer, set
the offset so that the drawer starts moving when the finger reaches
the drawer, not before.

Change-Id: Ibb8bf8f67f490e0706b6230080907bb185bd25bc
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>