Fix value changed logic for ValueTypeProvider types
Fix broken value changed logic for types implemented
through the ValueTypeProvider interface (QtQuick and Qt3d).
It affects vector2d, vector3d, ...

The signal was not emitted for the cases where the new value
was equal to the default value of a given type.

Also add a unit test to cover this area.

Change-Id: I9491b0462c78fecc4c704ea36921611c1bd6b2ee
Task-number: QTBUG-50204
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>