StackView: fix stack clear on replace of the topmost item
On item replace, StackView pops the topmost item (to use in transition)
and call QQuickStackViewPrivate::replaceElements(), which delete items
until it find the target item. When replacing the topmost item, it just
clear the whole stack, as the target item not in the stack already and
it procced to the bottom of the stack.

This commit changes StackView to use pushElements() for replacing the
topmost item, as there is nothing to replace.

Task-number: QTBUG-50274
Change-Id: Iff2e1177ee2f8b2d4518b519b450ae4e89a4c963
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>