qmlplugindump: avoid generating conflicting types.
qmlplugindump used to import QtQuick, therefore skipping all QtQuick
types when dumping.  Now that it imports only Qt, it is no longer the
case, and all QtQuick types would be dumped in the module's qmltypes
file.  It can be avoided by specifyng 'QtQuick' as dependency in qmldir
file or in a json file passed as -dependencies option.  It seems however
that people are not used to that, so in order to restore the old
behavior, the QtQuick dependency is automatically added when the
dependency list is empty.

Change-Id: I5fb2e57893607a877d284767b3fd09159b45ff42
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@theqtcompany.com>