qmlplugindump: option to merge qmltypes.
QtCreator does not handle dependencies between Qml types defined in
different qmltypes files.  Sometimes manual editing of qmltypes file is
needed to let QtCreator find the missing type information.
With the new -merge option it is possible to merge a qmltypes file to
the output of qmlplugindump.  Dependencies are correctly merged but
components are simply added, so they could cause conflict.

Change-Id: I6569339e4f05d37ea63fa2173983b4d595ae0ad6
Reviewed-by: Thomas Hartmann <Thomas.Hartmann@theqtcompany.com>