QSGDefaultRectangleNode: Fix off-by-one color rendering
Implicit conversion from floating-point to int, which simply truncates
the number's fractional part, may have caused the rectangle fill
to be rendered in a color with some of its RGB components different
by ±1 from the specified value.

For example, the actual on-screen color of the following QML item was
#010101 on 32bit Intel (no antialiasing or alpha channel):

  Rectangle {
      width: 100
      height: 100
      color: '#020202'
  }

This commit improves precision of the calculation by applying qRound()
before the conversion to integer.

Change-Id: Ia849c31ba9872e7a92608245406d178051b21917
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>