Restore Material::primary
The idea was that users would simply configure a different accent color
for the toolbar, but the only way to do this is in code, which creates
a hard dependency to the material style. A separate primary color makes
it possible to configure both colors in :/qtlabscontrols.conf, which
nicely avoids style specific dependencies.

Change-Id: Ia5ad92b57f9aea32e16e9260117e8d32b3356556
Task-number: QTBUG-50329
Reviewed-by: Nikita Krupenko <krnekit@gmail.com>
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>