Fix copied QDragMoveEvent drop action to propagate to original event
Trying to set the drop action and the accepted state in a overridden
dragMoveEvent handler, does not get propagated to the original
QDragMoveEvent, because the event passed to the handler is a copy.

This does not allow canceling the drop action in the move handler,
or change the proposed action to a different one. Changing these
values in the move handler is important to allow modifying the cursor
when moving / hovering above a possible drop item, depending on
user conditions.

Fix consists in copying the drop action and accepted values to the
original event, as well as removing the hard-coded setAccepted
(true) call.

Task-number: QTBUG-58260
Change-Id: I7a4bd4e68ee1023a36a63d3e835c282077e4187c
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>