Add QQuickPopup::popupItem()
Even though I didn't want to expose this, because I'm still hoping we
could make QQuickPopup inherit QQuickControl in the future, the fact
that QQuickPopupPrivate::get(popup)->popupItem was used in so many
places is a good indicator that accessing the popup item is often
required when dealing with popups. In any case, this is C++ only API
and not exposed to QML, so it's not that big deal... and we can
always deprecate it later and make it return "this" should the item
plans come true.

Change-Id: I6b99a499327d838ee61eae70f8ebf8e77f00ae39
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>