Add Popup::margins
Allow specifying margins around the popup that are taken into account
while positioning the popup. For example, the Material style draws a
drop shadow that goes outside the popups bounds. This allows it to
specify margins so that the drop shadow doesn't get clipped. Other
use case is to specify the margins so that the popup doesn't overlap
with window header or footer.

Change-Id: I018d961259fa8b960d86b77de01413eefb75a59b
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>