Make popups work without ApplicationWindow, to some degree
Using ApplicationWindow is highly recommended. First of all, with a
plain Window, QQuickPopup attempts to set a high z-value, but cannot
guarantee correct stacking order. Secondly, we cannot provide style-
specific background dimming for modal popups, because it is styled
as part of ApplicationWindow. Last but not least, QQuickPopup has to
install a window-level event filter, which is far less efficient than
how event handling done in QQuickOverlay.

Change-Id: I08915abce7a1764177b92f7539eef77c054a405a
Task-number: QTBUG-49921
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>