Fix off-by-one when checking overlaps in selection ranges
This was found while researching QTBUG-49596. The selection
ranges in the QGlyphRun are inclusive, so the length needs
to be end - start + 1. For a node of 1 glyph with either
preceding or succeeding node, we would detect an overlap
always, since the initial rangeLength would be 0. This
is reproduced by the textinput_selected_fallback_font.qml
test, but the bug was hidden by a different bug in
QTextLayout.

Change-Id: I65d70b1223eebeb5cfbb277fade7f4753465364f
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Simon Hausmann <simon.hausmann@theqtcompany.com>