Include qqml.h in qqmlengine.h for convenience
When a user creates a Qt Quick Application Project in Creator, they get
this main.cpp:

int main(int argc, char *argv[])
{
    QGuiApplication app(argc, argv);

    QQmlApplicationEngine engine;
    engine.load(QUrl(QStringLiteral("qrc:/main.qml")));

    return app.exec();
}

According to the current documentation, if they then want to register a
C++ type for use in QML, they need to add a separate
"#include <QtQml>" (which is actually a module-wide include and brings
a lot of unnecessary files with it).

If the most common (or perhaps only) use case for that function (and
all of the others declared in qqml.h but documented as requiring QtQml)
is related to QQmlEngine, then it makes sense for qqmlengine.h to
include qqml.h.

This change makes it so that users can use these functions with
no changes to the code that Creator provides, as QQmlApplicationEngine
includes QQmlEngine.

Change-Id: I9d6be2e1abfd4cc831ca04dab4c2a9e50021f748
Task-number: QTBUG-50343
Reviewed-by: J-P Nurmi <jpnurmi@theqtcompany.com>
Reviewed-by: Lars Knoll <lars.knoll@theqtcompany.com>