Remove QQuickFrame::frame in favor of background
A separate frame property was invented back in the days when we still
planned to have only one style, and when the Pane control still didn't
exist. Now there are multiple styles to choose from, each drawing a
different style frame, so it is no longer that important to be able
to have so fine-grained delegate API. Also, instead of doing
Frame.frame.visible:false one can now conveniently use Pane for a
frameless container. Furthermore, "frame" as a property name was prone
to conflict with QML element identifiers in applications.

Change-Id: Ia6465526e1f393884ac84cb56625cc080decd2af
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>