Build (and fail) Quick with -no-opengl
Disable qmltest, particles, Quick examples and tests.

Also guard private QOpenGL header includes because, unlike the public
QOpenGL* headers, this is not done in the private headers.

This allows build attempts against a -no-opengl qtbase. It will of course
fail when encontering the GL symbols and QOpenGL class usage in Quick/SG.

Change-Id: I323f7a8b8854cc6e486e264517df30be3e5f8d3c
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>