Remove AbstractButton::label in favor of contentItem
This has been on the TODO list for a long time and should have been
done earlier. It's time to cleanup the API and remove all the delegate
properties that became redundant when Control::contentItem was
introduced.

The biggest benefit of using Control::contentItem is that it is resized
and positioned automatically, so we save a few QML bindings. Furthemore,
contentItem is a common naming pattern in Qt Quick (Window, Flickable)
and therefore less likely to conflict with IDs in user code (a thing
that always bothered me with "label").

Change-Id: I8143a9f08d726db31cf4618670af48bd37f80b89
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>