Fix Animators in itemview and positioner transitions
In itemview/positioner transitions, Animator's target-property is never
explicitly set. QQuickItemViewTransition passes the animation target as
a "default target" instead. This change adds missing handling for the
default target.

QQuickItemViewTransition sets up default "x" and "y" state actions.
If Animator drives opacity, scale, or basically anything else than x/y
animations, it failed to extract the from/to values from the state
action list.

This change fixes the issue that if the default state actions do not
match the animator property (x/y vs. scale/opacity), it uses from/to
values specified on the animator itself. Before, it did that only if
the default state action list was empty. This is not the case with
itemview/positioner transitions.

Change-Id: I0f15e20bc860ddec23e59efebbc9cd346317f4de
Task-number: QTBUG-50908
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>
Reviewed-by: Robin Burchell <robin.burchell@viroteck.net>