Make it possible to have built-in adaptations
In addition to QMLSCENE_DEVICE, start supporting an env var named
QT_QUICK_BACKEND. This is more suitable for general use.

Besides plugin-based adaptations, some built-in ones are anticipated as
well. Therefore, add support for shipping custom contexts as part of the
scenegraph. For test purposes, a dummy adaptation is provided that falls
back to the default, OpenGL-based path for everything. This may
be removed at some point later on, but is useful for testing the loading
logic for now.

Change-Id: I7e0a2cd3ef33ba910604fb4027af3b3711fcaf1e
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>