Speedup tst_pressandhold
Before, running the test took 8032ms. After the patch it takes 1375ms.

The test has been implemented without timers. It creates one control
instance for testing signal emission/cancellation, and another just for
waiting for the signal. By the time the second control emits, the first
control has had to emit or cancel the signal. Thus, for the test, the
exact press and hold interval does not matter.

Change-Id: I3016217c2422cb8d79e56b0b51a11eaeef08865a
Reviewed-by: Mitch Curtis <mitch.curtis@theqtcompany.com>