Clear profiling data from QQuickProfiler after reporting
This is what 5.6 does. It's debatable if this is actually correct,
as logically we should report the same global profiling data multiple
times, when data for different engines is reported. However, there
is no practical benefit for such a "feature". It just wastes CPU time
and memory and it makes the client implementation more complicated.

Change-Id: I6d2a253d906d3d74c9d9b1f8fc952673c9da3e2a
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>