Popup: use available size and paddings from the popup item
Now that the internal popup item inherits Control, which already
provides these features, we may use them directly instead of re-doing
them in Popup.

NOTE: QQuickMenu code has lots of direct references to the content
item, which used to be stored in QQuickPopupPrivate, which is the
base class of QQuickMenuPrivate. Now that the content item is stored
by QQuickPopupItem instead, the direct references don't work. Thus,
as a temporary solution to avoid this patch growing large, we'll
store the content item in QQuickMenuPrivate.

Change-Id: I80a16d34432131a4326c1b6ae54419470be53f96
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>