Flickable: add AutoFlickIfNeeded as an option for flickableDirection
If you want to disable all movement when the user tries to flick or
drag, you previously had to set interactive: false.  Maybe you need to
bind that to some calculation to determine whether the content fits
completely inside the Flickable or not.  This way is easier.

BTW the AutoFlickIfNeeded can be ORed with HorizontalFlick or
VerticalFlick, but we don't document it because AutoFlickDirection=0,
so it's not useful to OR that with anything.

Task-number: QTBUG-31121
Change-Id: Ib03b0f223cb40f0338510c318aa37e70ce71514d
Reviewed-by: Jan Arve Sæther <jan-arve.saether@theqtcompany.com>