D3D12: add opacity support, fix buffer handling and dirty tracking
Avoid using .h for generated headers with shader blobs in order not to
confuse syncqt.

All buffer Map and memcpy is now done only at the end of the frame. This
is relevant for the constant buffer which can this way Map once and memcpy
only the data for elements that changed something.

This also marks a change from how simplerenderer works: instead of having
per-program matrix, opacity, etc. values (uniforms), there is a separate
copy of these in a dedicated area of the constant buffer for each element.
This allows reusing the data in the buffer as-is for non-changing elements
at the cost of higher memory usage (256 bytes per element both on CPU &
GPU side at minimum) and is also necessary in order to be able to build
command lists with multiple draw calls without intra-dependencies for the
cbuffer contents.

The somewhat broken dirty tracking of simplerenderer (matrices always
treated as dirty, forgetting the root node for ever in the list of dirty
transform nodes, etc.) is now corrected so that the DirtyMatrix and
DirtyOpacity states reflect reality, meaning the cbuffer update can be
skipped completely for items whose transform or inherited opacity has not
changed at all.

Format mapping is now moved to the engine.

Change-Id: Ia0a0d73e2a954970cb2745deb1a62c09f3d3cb7c
Reviewed-by: Andy Nichols <andy.nichols@theqtcompany.com>