Flickable: handle wheel NoScrollPhase the same as ScrollUpdate
Before change f253f4c in qtbase, when an ordinary mouse wheel is
rotated the phase would be ScrollUpdate continuously, without going
through the other phases; whereas on a gesture-interpreting trackpad
there would be a ScrollBegin first.  Now, an ordinary wheel mouse
uses NoScrollPhase continuously, which makes it easier to tell the
two cases apart.  But we're required to handle the new phase
explicitly.  To avoid change in existing behavior, we handle it
the same as ScrollUpdate phase was handled before.

Change-Id: I2d51a89a965704cf843303d523463805b4fe6ff8
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
Reviewed-by: Gabriel de Dietrich <gabriel.dedietrich@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>