tst_combobox: temporarily disable tests with two windows
The data-driven tests with both Window and ApplicationWindow added
in 068fe79 mean in practice that two windows are hanging around during
the tests. Even though we always try to activate the currently tested
window, the CI has troubles with these tests. These tests probably need
to be moved to C++ so that we have better control over which window is
shown and when.

Change-Id: I48839d590a72b12437b6eb6259296232802c8871
Reviewed-by: Liang Qi <liang.qi@theqtcompany.com>