QQuickTextNodeEngine: Update boundingRects when merging nodes
QGlyphRun has the behavior that it will calculate it's boundingRect
based on the united rects of its contained glyphs unless a boundingRect
has been previously set.  In the case of QQuickTextNodeEngine the
QGlphyRuns are coming from lines in a QTextDocument which would have
already had setBoundingRect called. For the QGlyphRun that results from
the merging of text nodes to have the correct boundingRect we would need
set the united boundingRects of the nodes merged.

This could go overlooked with the default renderer because it doesn't
need to know the boundingRect of the QGlyphRun to render accurately.
However this detail is important for partial update support with Qt
Quick 2D Renderer.

Change-Id: I4a5bd20afcfadf6cb0ad0659f9a797a980b47a81
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@theqtcompany.com>